// interfaces/worker.ts
export interface WorkerModelInterface {
  _id: string;
  // idCreador: any;
  nombre: string;
  apellido: string;
  correo: string;
  password: string;
  telefono: string;
  fecha_alta: string;
  fecha_login: string;
  role: string;
  cantVisitas: number;
  estado: boolean;
  sucursal: any;
  identificacion: string;
  foranea: any;
  empresa: boolean;
}